requirements.txt
astropy~=7.0
lmfit~=1.3
joblib~=1.4
matplotlib~=3.10
numpy~=2.2
pandas~=2.2
scipy~=1.15
scikit-learn~=1.6
tomli >= 2.0.0
bokeh~=3.6
mplcursors~=0.6
jax~=0.4
jaxlib~=0.4
pymc~=5.1
pytensor~=2.20
openpyxl~=3.1
PyLaTeX~=1.4
streamlit_bokeh==3.6.2
lime-stable==2.0.dev1
aspect-stable==0.3.0
lime-stable==2.0.dev1
innate-stable==0.2.0
specsy==0.6.0
